# packages/contracts/openapi/core-api.yaml
openapi: 3.1.0
info:
  title: Lab Management Platform Core API
  version: 0.1.0
  description: 核心宿主与插件公开 HTTP 契约。跨模块调用必须依赖此类契约，不得依赖实现代码或数据库。
servers:
  - url: http://localhost:3000
paths:
  /health:
    get:
      summary: 平台健康检查与已加载插件列表
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [status, plugins]
                properties:
                  status:
                    type: string
                    example: ok
                  plugins:
                    type: array
                    items:
                      type: string
  /plugins/hello-world:
    get:
      summary: HelloWorld 示例插件
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message:
                    type: string
  /auth/login:
    post:
      summary: 本地账号、学号/工号或手机号 + 密码登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  description: 可填写登录账号、学号/工号或已绑定手机号
                password:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [token, actor]
                properties:
                  token:
                    type: string
                  actor:
                    $ref: "#/components/schemas/Actor"
        "401":
          description: 用户名或密码错误
  /auth/register:
    post:
      summary: 管理员创建本地账号
      description: 当前用于学号/工号与密码注册；后续可由统一身份认证适配器替换或禁用。
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, studentId, displayName, role]
              properties:
                username:
                  type: string
                password:
                  type: string
                  minLength: 8
                studentId:
                  type: string
                  pattern: "^[A-Za-z0-9_-]{4,32}$"
                displayName:
                  type: string
                role:
                  type: string
                  enum: [admin, member]
      responses:
        "201":
          description: 已创建账号
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Actor"
        "403":
          description: 无注册权限
        "409":
          description: 用户名或学号/工号已存在
  /auth/profile:
    get:
      summary: 获取当前用户资料
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedUser"
  /auth/profile/contact:
    patch:
      summary: 当前用户绑定或修改手机号
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone:
                  type: string
                  pattern: "^1[3-9]\\d{9}$"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedUser"
        "409":
          description: 手机号已被其他账号绑定
  /auth/profile/password:
    patch:
      summary: 当前用户修改本地登录密码
      description: 统一身份认证账号后续应由认证提供方处理密码修改。
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
                  minLength: 8
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
        "403":
          description: 当前密码错误
  /auth/users:
    get:
      summary: 管理员查询账号列表
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: 按账号、姓名、学号/工号、手机号模糊搜索
        - name: includeInactive
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: 是否包含已停用账号；默认只返回启用账号
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ManagedUser"
        "403":
          description: 无账号查看权限
  /auth/users/{id}/password:
    patch:
      summary: 管理员重置学员密码
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newPassword]
              properties:
                newPassword:
                  type: string
                  minLength: 8
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
        "403":
          description: 无重置权限
  /auth/users/{id}/role:
    patch:
      summary: 管理员调整用户角色
      description: 可在 member 与 admin 之间调整；不能调整当前登录账号与 super_admin。
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [admin, member]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedUser"
        "403":
          description: 无角色管理权限
  /auth/users/{id}:
    delete:
      summary: 管理员删除学员账号
      description: 软删除学员账号，保留历史业务记录；管理员账号不能通过此接口删除。
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
        "403":
          description: 无删除权限
  /inventory/applications:
    get:
      summary: 获取耗材申请列表
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryApplication"
    post:
      summary: 提交耗材或设备申请
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [materialId, quantity]
              properties:
                materialId:
                  type: string
                quantity:
                  type: integer
                  minimum: 1
      responses:
        "202":
          description: 申请已接收，后续由审批流处理
          content:
            application/json:
              schema:
                type: object
                required: [applicationId, status]
                properties:
                  applicationId:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum: [pending]
        "401":
          description: 未认证
        "400":
          description: 请求数据不合法
  /inventory/summary:
    get:
      summary: 获取耗材管理统计
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [materialCount, lowStockCount, pendingApplications, approvedApplications]
                properties:
                  materialCount:
                    type: integer
                  lowStockCount:
                    type: integer
                  pendingApplications:
                    type: integer
                  approvedApplications:
                    type: integer
  /inventory/materials:
    get:
      summary: 获取耗材库存列表
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Material"
  /inventory/stock-movements:
    get:
      summary: 查询库存流水
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockMovement"
  /files:
    get:
      summary: 查询文件资料
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: 按标题、分类、说明模糊搜索
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LabFile"
    post:
      summary: 登记 Synology Drive 文件资料
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, category, driveUrl]
              properties:
                title:
                  type: string
                category:
                  type: string
                  enum: [sop, template, record, other]
                driveUrl:
                  type: string
                  format: uri
                description:
                  type: string
      responses:
        "201":
          description: 已登记文件资料
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabFile"
        "403":
          description: 无文件管理权限
  /inventory/applications/{id}/approve:
    patch:
      summary: 管理员批准耗材申请
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                remark:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryApplication"
  /inventory/applications/{id}/reject:
    patch:
      summary: 管理员拒绝耗材申请
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                remark:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryApplication"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Actor:
      type: object
      required: [id, username, displayName, role, permissions]
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [super_admin, admin, member]
        permissions:
          type: array
          items:
            type: string
    ManagedUser:
      type: object
      required: [id, username, displayName, role, identityProvider, active, createdAt]
      properties:
        id:
          type: string
        username:
          type: string
        studentId:
          type: string
        phone:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [super_admin, admin, member]
        identityProvider:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
    DomainEvent:
      type: object
      required: [id, type, version, occurredAt, source, payload]
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        version:
          type: integer
        occurredAt:
          type: string
          format: date-time
        source:
          type: string
        payload:
          type: object
    Material:
      type: object
      required: [id, name, spec, stock, warnStock, unit, location, manager]
      properties:
        id:
          type: string
        name:
          type: string
        spec:
          type: string
        stock:
          type: integer
        warnStock:
          type: integer
        unit:
          type: string
        location:
          type: string
        manager:
          type: string
    InventoryApplication:
      type: object
      required:
        [
          id,
          materialId,
          materialName,
          applicantId,
          applicantName,
          quantity,
          reason,
          status,
          createdAt
        ]
      properties:
        id:
          type: string
        materialId:
          type: string
        materialName:
          type: string
        applicantId:
          type: string
        applicantName:
          type: string
        quantity:
          type: integer
        reason:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
          format: date-time
        reviewedAt:
          type: string
          format: date-time
        reviewRemark:
          type: string
    StockMovement:
      type: object
      required: [id, materialId, materialName, operatorId, quantity, type, remark, createdAt]
      properties:
        id:
          type: string
        materialId:
          type: string
        materialName:
          type: string
        operatorId:
          type: string
        quantity:
          type: integer
        type:
          type: string
          enum: [stock_in, application_out]
        remark:
          type: string
        createdAt:
          type: string
          format: date-time
    LabFile:
      type: object
      required: [id, title, category, driveUrl, description, ownerId, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        category:
          type: string
          enum: [sop, template, record, other]
        driveUrl:
          type: string
          format: uri
        description:
          type: string
        ownerId:
          type: string
        createdAt:
          type: string
          format: date-time